Concussion-grade facepalming
As it turns out, I don't actually NEED to hide the keys for this
combination of services.  The error messages and issues I've been
struggling with were because I was trying to use a key where it wasn't
needed.
On the Programmer Competene Scale, I rate today as "too dumb to turn on
a computer".  But my project works now, so I've got that going for me,
which is nice.